pyproject.toml
[project]
name = "elon"
version = "0.1.0"
description = ""
authors = [
    {name = "Achyuth",email = "achyuthrayal2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "newspaper3k (>=0.2.8,<0.3.0)",
    "pymongo (>=4.16.0,<5.0.0)",
    "cloudscraper (>=1.2.71,<2.0.0)",
    "playwright (>=1.58.0,<2.0.0)",
    "youtube-transcript-api (>=1.2.4,<2.0.0)",
    "yt-dlp (>=2026.2.4,<2027.0.0)",
    "webvtt-py (>=0.5.1,<0.6.0)",
    "pandas (>=3.0.0,<4.0.0)",
    "pypdf (>=6.7.0,<7.0.0)",
    "sentence-transformers (>=5.2.2,<6.0.0)",
    "tqdm (>=4.67.3,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
